requirements.txt
fastapi
uvicorn
google-api-python-client
textblob
python-dotenv
pandas
scikit-learn
nltk
python-multipart
